LPS-47908 Move logic to more appropriate packages and define public api in bnd file
