+++ modules/util/markdown-converter/bnd.bnd
@@ -1,1 +1,3 @@
 Bundle-Name: Liferay Markdown Converter
+Export-Package: com.liferay.markdown.converter
+Private-Package: com.liferay.markdown.converter.internal.*
