+++ modules/util/markdown-converter/src/com/liferay/markdown/converter/internal/pegdown/ast/PicWithCaptionNode.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.pegdown.ast;
+package com.liferay.markdown.converter.internal.pegdown.ast;
 
 import java.util.List;
 
@@ -22,17 +22,14 @@ import org.pegdown.ast.TextNode;
 import org.pegdown.ast.Visitor;
 
 /**
- * Represents a "side-bar" image with associated text. A "side-bar" is currently
- * set apart from the surrounding text of a document by a horizontal rule above
- * and below the image and its associated text. The image is typically an icon
- * (e.g. a pen tip for a tip side-bar) that denotes the type of the "side-bar".
+ * Represents picture with an associated caption.
  *
  * @author James Hinkey
  */
-public class SidebarNode extends SuperNode {
+public class PicWithCaptionNode extends SuperNode {
 
-	public SidebarNode(Node captionNode, String src, Node node) {
-		super(captionNode);
+	public PicWithCaptionNode(String src, Node node) {
+		super(node);
 
 		_alt = _getAlt(node);
 		_src = src;
@@ -71,7 +68,7 @@ public class SidebarNode extends SuperNode {
 
 	private static final String _BLANK = "";
 
-	private String _alt;
-	private String _src;
+	private final String _alt;
+	private final String _src;
 
 }
