+++ modules/util/markdown-converter/src/com/liferay/markdown/converter/internal/pegdown/parser/LiferayParser.java
@@ -12,13 +12,15 @@
  * details.
  */
 
-package org.pegdown;
+package com.liferay.markdown.converter.internal.pegdown.parser;
 
-import com.liferay.pegdown.ast.PicWithCaptionNode;
-import com.liferay.pegdown.ast.SidebarNode;
+import com.liferay.markdown.converter.internal.pegdown.ast.PicWithCaptionNode;
+import com.liferay.markdown.converter.internal.pegdown.ast.SidebarNode;
 
 import org.parboiled.Rule;
 import org.parboiled.common.ArrayBuilder;
+
+import org.pegdown.Parser;
 
 /**
  * Provides rules for parsing pictures with captions, "side-bars", and in-line
